requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-cloud-firestore==2.13.1
google-generativeai==0.3.2
pywebpush==1.14.1 
markdown